+++ barapost/src/sorter_modules/launch.py
@@ -9,7 +9,7 @@ from src.spread_files_equally import spread_files_equally
 
 
 def launch_single_thread_sorting(fpath_list, str_func, tax_annot_res_dir, sens,
-    min_qual, min_qlen, logfile_path):
+    min_qual, min_qlen, min_pident, min_coverage, logfile_path):
     """
     Function launches single-thread sorting, performed by finction 'srt_func'.
 
@@ -20,10 +20,14 @@ def launch_single_thread_sorting(fpath_list, str_func, tax_annot_res_dir, sens,
     :type tax_annot_res_dir: str;
     :param sens: sorting sensitivity;
     :type sens: str;
-    :param min_qual: minimum quality to keep;
+    :param min_qual: threshold for quality filter;
     :type min_qual: float;
-    :param min_qlen: minimmum sequence length to keep;
-    :type min_qlen: int (or None, if this feature is disabled);
+    :param min_qlen: threshold for length filter;
+    :type min_qlen: int (or None, if this filter is disabled);
+    :param min_pident: threshold for alignment identity filter;
+    :type min_pident: float (or None, if this filter is disabled);
+    :param min_coverage: threshold for alignment coverage filter;
+    :type min_coverage: float (or None, if this filter is disabled);
     :param logfile_path: path to log file;
     :type logfile_path: str;
     """
@@ -31,7 +35,8 @@ def launch_single_thread_sorting(fpath_list, str_func, tax_annot_res_dir, sens,
     # Sort files in single thread:
     res_stats = map(
         partial(str_func, tax_annot_res_dir=tax_annot_res_dir, sens=sens,
-            min_qual=min_qual, min_qlen=min_qlen, logfile_path=logfile_path),
+            min_qual=min_qual, min_qlen=min_qlen, min_pident=min_pident,
+            min_coverage=min_coverage, logfile_path=logfile_path),
         fpath_list)
 
     return res_stats
@@ -39,7 +44,7 @@ def launch_single_thread_sorting(fpath_list, str_func, tax_annot_res_dir, sens,
 
 
 def launch_parallel_sorting(fpath_list, str_func, tax_annot_res_dir, sens, n_thr,
-    min_qual, min_qlen, logfile_path):
+    min_qual, min_qlen, min_pident, min_coverage, logfile_path):
     """
     Function launches single-thread sorting, performed by finction 'srt_func'.
 
@@ -52,10 +57,14 @@ def launch_parallel_sorting(fpath_list, str_func, tax_annot_res_dir, sens, n_thr
     :type sens: str;
     :param n_thr: number of threads to launch;
     :type n_thr: int;
-    :param min_qual: minimum quality to keep;
+    :param min_qual: threshold for quality filter;
     :type min_qual: float;
-    :param min_qlen: minimmum sequence length to keep;
-    :type min_qlen: int (or None, if this feature is disabled);
+    :param min_qlen: threshold for length filter;
+    :type min_qlen: int (or None, if this filter is disabled);
+    :param min_pident: threshold for alignment identity filter;
+    :type min_pident: float (or None, if this filter is disabled);
+    :param min_coverage: threshold for alignment coverage filter;
+    :type min_coverage: float (or None, if this filter is disabled);
     :param logfile_path: path to log file;
     :type logfile_path: str;
     """
@@ -75,6 +84,8 @@ def launch_parallel_sorting(fpath_list, str_func, tax_annot_res_dir, sens, n_thr
         n_thr=n_thr,
         min_qual=min_qual,
         min_qlen=min_qlen,
+        min_pident=min_pident,
+        min_coverage=min_coverage,
         logfile_path=logfile_path), [(sublist,) for sublist in spread_files_equally(fpath_list, n_thr)])
 
     pool.close()